New Callback Override System
In a nutshell: This now allows you to override AND monkey-patch the current callbacks based on the love.* references. However, copying/moving the referenced or monkey-patched versions of those callbacks to another position may cause issues. It's also not been tested too much, so if anything breaks, you know who to yell at.