Tweaked Callback Handling
SHOULD now allow having tables there. Trying to access overridden callbacks will now always give you a callable table. Doing so will result in calling the original or the callback of the debugger, depending on whether it is active...

I'm still not quite happy with this, but it at least covers another eventuality.